Add first Link uploading version.
This is very much a work in progress. At the moment there is a sinlge
user and in general even with that limited amount the UX is very rough.

Next is to make the UI better. I think I will use GitHubs Primer
Framework so the UI will feel native to the developers.